requirements.txt
mcp>=1.0.0
anthropic>=0.40.0
httpx>=0.27.0
pandas>=2.0.0
pandas-ta>=0.3.14b
numpy>=1.26.0
python-dotenv>=1.0.0
